+++ include/linux/mlx4/device.h
@@ -536,7 +536,7 @@ struct mlx4_caps {
 	int			comp_pool;
 	int			num_mpts;
 	int			max_fmr_maps;
-	int			num_mtts;
+	u64			num_mtts;
 	int			fmr_reserved_mtts;
 	int			reserved_mtts;
 	int			reserved_mrws;
